Function to identify STR listings which are likely to be principal residences

strr_principal_res takes a set of inputs about listing activity and identifies listings which are likely to be principal residences in a given time frame.

strr_principal_res(
  property,
  daily,
  host,
  FREH,
  ghost,
  start_date = NULL,
  end_date = NULL,
  tolerance = 0,
  field_name = principal_res,
  quiet = FALSE
)

Arguments

property
A data frame of STR listings in standard UPGo format.

daily
A data frame of daily STR activity in standard UPGo format.

host
A data frame of STR host activity in standard UPGo format.

FREH
A data frame with frequently rented entire-home listing status by day, of the type produced with strr_FREH.

ghost
A data frame with ghost hostel activity by day, of the type produced with strr_ghost

start_date
A character string of format YYYY-MM-DD indicating the first date for which to assess principal residence status. If NULL (default), the earliest date common to the `daily`, `host`, `FREH` and `ghost` tables will be used.

end_date
A character string of format YYYY-MM-DD indicating the last date for which to assess principal residence status. If NULL (default), the latest date common to the `daily`, `host`, `FREH` and `ghost` tables will be used.

tolerance
A number between 0 (default) and 1 setting the tolerance with which daily multilisting, FREH and ghost hostel statuses should be converted into a single assessment of these statuses across the entire date range. A value of, e.g., 0.1 means that if a listing has FREH status on more than 10 the purpose of setting principal residence status. A value of 1 means that the listing must be FREH on every single date in the date range to be considered FREH for the purpose of setting principal residence status, while a value of 0 means that if the listing is FREH on any number of dates greater than 0 then it will be considered FREH for the purpose of setting principal residence status.

field_name
A character string to name the new logical field identifying principal residences in the output table. The default is `principal_res`.

quiet
A logical vector. Should the function execute quietly, or should it return status updates throughout the function (default)?

Value

The output will be the `property` input table with one additional logical field (with name taken from the `field_name` argument) indicating principal residence status.

Details

TKTK explanation